refactor: make `ReadDetect` into a resource
This is to show the pattern I always use for threads.

`ReadDetect` represents a thread/worker. Creating it
starts it, eliminating the need for the caller to bother handling
anything about it.

The returned value is a real handle. It will join automatically
on a drop (methods could be added to change that behavior if desired).

There's a little bit of right drift as the `start` was inlined,
but it could be refactored out into a static function if desired.